Create Message_Testing.ino
Just adding the code I used to monitor the functionality of the Meshtastic while on a Balloon flight. This has the same functionality of the data_receiver.ino code, but there is code to send an increasing number (1-20) every 20 seconds to show you are getting new messages. Tested in the field with great success, far beyond expectations.